Build and publish to Docker Hub with CCC
This repo is a fork of:

- https://github.com/wallneradam/docker-coreos-toolbox-ubuntu

We're going to use this on the kubectl instances so that we have tmux
available for pairing on incidents. We want to build our own copy of it
because:

- we don't want to depend on a third-party image in case it disappears
  in future
- we want to make some modifications to it such as making kubectl
  available within the container

We're using CCC to publish it because it gives us Twistlock scanning for
free. We're only pushing to a public Docker Hub repo, not ECR, because
the kubectl instance doesn't have any credentials to pull private repos.